Add run date to list
