+++ src/components/PipelineRow/index.tsx
@@ -1,5 +1,4 @@
 import { Link, useNavigate } from "@tanstack/react-router";
-import localForage from "localforage";
 import { List } from "lucide-react";
 import { useEffect, useState } from "react";
 
@@ -10,16 +9,13 @@ import type {
 import { downloadDataWithCache, loadObjectFromYamlData } from "@/cacheUtils";
 import { TableCell, TableRow } from "@/components/ui/table";
 import { API_URL, EDITOR_PATH } from "@/utils/constants";
+import { fetchPipelineRuns, type PipelineRun } from "@/utils/fetchPipelineRuns";
 
 import { Popover, PopoverContent, PopoverTrigger } from "../ui/popover";
 import { ScrollArea } from "../ui/scroll-area";
 import RunListItem from "./RunListItem";
 import StatusIcon from "./StatusIcon";
-import {
-  type PipelineRowProps,
-  type PipelineRun,
-  type TaskStatusCounts,
-} from "./types";
+import { type PipelineRowProps, type TaskStatusCounts } from "./types";
 import { countTaskStatuses, formatDate, getRunStatus } from "./utils";
 
 const PipelineRow = ({ url, componentRef, name }: PipelineRowProps) => {
@@ -63,43 +59,18 @@ const PipelineRow = ({ url, componentRef, name }: PipelineRowProps) => {
   }, [url, componentRef, name]);
 
   useEffect(() => {
-    const fetchPipelineRuns = async () => {
+    const fetchData = async () => {
       if (!name) return;
 
-      try {
-        const pipelineRunsDb = localForage.createInstance({
-          name: "components",
-          storeName: "pipeline_runs",
-        });
+      const res = await fetchPipelineRuns(name);
 
-        const runs: PipelineRun[] = [];
-        let latestRun: PipelineRun | null = null;
-        let latestDate = new Date(0);
+      if (!res) return;
 
-        await pipelineRunsDb.iterate<PipelineRun, void>((run) => {
-          if (run.pipeline_name === name) {
-            runs.push(run);
-            const runDate = new Date(run.created_at);
-            if (runDate > latestDate) {
-              latestDate = runDate;
-              latestRun = run;
-            }
-          }
-        });
-
-        runs.sort(
-          (a, b) =>
-            new Date(b.created_at).getTime() - new Date(a.created_at).getTime(),
-        );
-
-        setPipelineRuns(runs);
-        setLatestRun(latestRun);
-      } catch (error) {
-        console.error("Error fetching pipeline runs:", error);
-      }
+      setPipelineRuns(res.runs);
+      setLatestRun(res.latestRun);
     };
 
-    fetchPipelineRuns();
+    fetchData();
   }, [name]);
 
   useEffect(() => {
