+++ src/components/PipelineRow/types.ts
@@ -1,12 +1,3 @@
-export interface PipelineRun {
-  id: number;
-  root_execution_id: number;
-  created_at: string;
-  pipeline_name: string;
-  pipeline_digest?: string;
-  status?: string;
-}
-
 export interface TaskStatusCounts {
   total: number;
   succeeded: number;
